+++ CHANGELOG.md
@@ -4,4 +4,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
 ### Added
--`JSONKtor`
+- `JSONKtor`: JSON content converter for ktor